requirements.txt
anyio==3.6.2
asyncpg==0.27.0
binance-connector>=2.0,<3.0
click==8.1.3
fastapi==0.94.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2>=3.1.2,<4.0
loguru>=0.6.0
MarkupSafe==2.1.2
numpy==1.24.2
pandas>=1.5.3,<2.0
psycopg-binary==3.1.8
psycopg-pool==3.1.6
pydantic==1.10.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy>=1.4.40,<2.0
starlette==0.26.0.post1
typing_extensions==4.5.0
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
